requirement.txt
fastapi
uvicorn
"mcp[cli]"
"langchain[anthropic]"
langchain-mcp-adapters
langgraph
httpx
pydantic
python-dotenv
